reporting: move rendering logic into `render.rs`
