+++ compiler/hash-reporting/src/builder.rs
@@ -1,8 +1,6 @@
 //! Hash compiler diagnostic report builder.
-
-use hash_error_codes::error_codes::HashErrorCode;
-
 use crate::report::{Report, ReportElement, ReportKind};
+use hash_error_codes::error_codes::HashErrorCode;
 
 /// A utility struct that allows for a [Report] to be built incrementally
 /// adding annotations and other metadata to the report.
