+++ compiler/hash-reporting/src/lib.rs
@@ -2,5 +2,6 @@
 pub mod builder;
 pub mod errors;
 pub mod highlight;
+mod render;
 pub mod report;
 pub mod writer;
