+++ compiler/hash-reporting/src/report.rs
@@ -1,17 +1,24 @@
-use std::{
-    cell::Cell,
-    fmt,
-    iter::{once, repeat},
-};
+//! Hash diagnostic report data structures.
+use std::{cell::Cell, fmt};
 
+use crate::highlight::{highlight, Colour, Modifier};
 use hash_error_codes::error_codes::HashErrorCode;
-use hash_source::{location::SourceLocation, SourceMap};
-use hash_utils::path::adjust_canonicalization;
+use hash_source::location::SourceLocation;
 
-use crate::{
-    highlight::{highlight, Colour, Modifier},
-    writer::{offset_col_row, ReportCodeBlockInfo, ERROR_MARKING_CHAR},
-};
+/// A data type representing a comment/message on a specific span in a code block.
+#[derive(Debug, Copy, Clone, Eq, PartialEq)]
+pub struct ReportCodeBlockInfo {
+    /// How many characters should be used for line numbers on the side.
+    pub indent_width: usize,
+    /// The beginning column of the code block.
+    pub start_col: usize,
+    /// The beginning row of the code block.
+    pub start_row: usize,
+    /// The end column of the code block.
+    pub end_col: usize,
+    /// The end row of the code block.
+    pub end_row: usize,
+}
 
 /// Enumeration describing the kind of [Report]; either being a warning, info or an
 /// error.
@@ -89,19 +96,6 @@ impl ReportNote {
             message: message.to_string(),
         }
     }
-
-    fn render(&self, f: &mut fmt::Formatter<'_>, longest_indent_width: usize) -> fmt::Result {
-        writeln!(
-            f,
-            "{} {} {}: {}",
-            " ".repeat(longest_indent_width),
-            highlight(Colour::Blue, "="),
-            highlight(Modifier::Bold, &self.label),
-            self.message
-        )?;
-
-        Ok(())
-    }
 }
 
 /// Data structure representing an associated block of code with a report. The type
@@ -123,155 +117,6 @@ impl ReportCodeBlock {
             info: Cell::new(None),
         }
     }
-
-    // Get the indent widths of this code block as (outer, inner).
-    pub(crate) fn info<T: SourceMap>(&self, modules: &T) -> ReportCodeBlockInfo {
-        match self.info.get() {
-            Some(info) => info,
-            None => {
-                let source_id = self.source_location.source_id;
-                let source = modules.contents_by_id(source_id);
-
-                let location = self.source_location.span;
-
-                let (start_col, start_row) = offset_col_row(location.start(), source, true);
-                let (end_col, end_row) = offset_col_row(location.end(), source, false);
-                let (_, last_row) = offset_col_row(source.len(), source, false);
-
-                let (top_buf, bottom_buf) = compute_buffers(start_row, end_row);
-                let indent_width = (start_row.saturating_sub(top_buf) + 1)
-                    .max((end_row + bottom_buf).min(last_row) + 1)
-                    .to_string()
-                    .chars()
-                    .count();
-
-                let info = ReportCodeBlockInfo {
-                    indent_width,
-                    start_col,
-                    start_row,
-                    end_col,
-                    end_row,
-                };
-
-                self.info.replace(Some(info));
-                info
-            }
-        }
-    }
-
-    /// Function to render the [ReportCodeBlock] using the provided [SourceLocation], message and
-    /// [ReportKind].
-    fn render<T: SourceMap>(
-        &self,
-        f: &mut fmt::Formatter,
-        modules: &T,
-        longest_indent_width: usize,
-        report_kind: ReportKind,
-    ) -> fmt::Result {
-        // Print line numbers, separator, then the code itself, split by newlines
-        // Should put the code_message where it is supposed to be with a caret (^) followed by (_)
-        // or equivalent unicode character.
-        //
-        // Print a few lines before and after location/code_message.
-        let source_id = self.source_location.source_id;
-        let source = modules.contents_by_id(source_id);
-        let ReportCodeBlockInfo {
-            start_row,
-            end_row,
-            start_col,
-            end_col,
-            ..
-        } = self.info(modules);
-
-        let (top_buffer, bottom_buffer) = compute_buffers(start_row, end_row);
-
-        let error_view = source
-            .trim_end_matches('\n')
-            .split('\n')
-            .enumerate()
-            .skip(start_row.saturating_sub(top_buffer))
-            .take(top_buffer + end_row - start_row + 1 + bottom_buffer);
-
-        // Print the filename of the code block...
-        writeln!(
-            f,
-            "{}{} {}",
-            " ".repeat(longest_indent_width),
-            highlight(Colour::Blue, "-->"),
-            highlight(
-                Modifier::Underline,
-                format!(
-                    "{}:{}:{}",
-                    adjust_canonicalization(modules.path_by_id(source_id)),
-                    start_row + 1,
-                    start_col + 1,
-                )
-            )
-        )?;
-
-        // Print each selected line with the line number
-        for (index, line) in error_view {
-            let index_str = format!(
-                "{:>pad_width$}",
-                index + 1,
-                pad_width = longest_indent_width
-            );
-
-            writeln!(
-                f,
-                "{} {}   {}",
-                if (start_row..=end_row).contains(&index) && !line.is_empty() {
-                    highlight(report_kind.as_colour(), &index_str)
-                } else {
-                    index_str
-                },
-                highlight(Colour::Blue, "|"),
-                line
-            )?;
-
-            if (start_row..=end_row).contains(&index) && !line.is_empty() {
-                let dashes = repeat(ERROR_MARKING_CHAR).take(
-                    if index == start_row && start_row == end_row {
-                        end_col - start_col
-                    } else if index == start_row {
-                        line.len().saturating_sub(start_col)
-                    } else if index == end_row {
-                        end_col
-                    } else {
-                        line.len()
-                    },
-                );
-
-                let mut code_note: String = repeat(" ")
-                    .take(if index == start_row { start_col } else { 0 })
-                    .chain(dashes)
-                    .collect();
-
-                if index == end_row {
-                    code_note.extend(once(" ").chain(once(self.code_message.as_str())));
-                }
-
-                writeln!(
-                    f,
-                    "{} {}   {}",
-                    " ".repeat(longest_indent_width),
-                    highlight(Colour::Blue, "|"),
-                    highlight(report_kind.as_colour(), &code_note)
-                )?;
-            }
-        }
-
-        Ok(())
-    }
-}
-
-/// Compute the top buffer and bottom buffer sizes from the given start and
-/// end row for a code block.
-fn compute_buffers(start_row: usize, end_row: usize) -> (usize, usize) {
-    const MAX_BUFFER: usize = 5;
-    let top_buffer: usize = (end_row - start_row + 1).min(MAX_BUFFER);
-    let bottom_buffer: usize = (end_row - start_row + 1).min(MAX_BUFFER);
-    (top_buffer, bottom_buffer)
 }
 
 /// Enumeration representing types of components of a [Report]. A [Report] can be made of
@@ -280,23 +125,6 @@ fn compute_buffers(start_row: usize, end_row: usize) -> (usize, usize) {
 pub enum ReportElement {
     CodeBlock(ReportCodeBlock),
     Note(ReportNote),
-}
-
-impl ReportElement {
-    pub(crate) fn render<T: SourceMap>(
-        &self,
-        f: &mut fmt::Formatter,
-        modules: &T,
-        longest_indent_width: usize,
-        report_kind: ReportKind,
-    ) -> fmt::Result {
-        match self {
-            ReportElement::CodeBlock(code_block) => {
-                code_block.render(f, modules, longest_indent_width, report_kind)
-            }
-            ReportElement::Note(note) => note.render(f, longest_indent_width),
-        }
-    }
 }
 
 /// The report data type represents the entire report which might contain many [ReportElement]s. The
